+++ sql/index.js
@@ -14,6 +14,8 @@ var fs = require('fs');
  * @type {{field: *, constraint: *, index: *}}
  */
 module.exports = {
+    schema      : fs.readFileSync(__dirname + '/schema.sql').toString(),
+    table       : fs.readFileSync(__dirname + '/table.sql').toString(),
     column      : fs.readFileSync(__dirname + '/column.sql').toString(),
     constraint  : fs.readFileSync(__dirname + '/constraint.sql').toString(),
     foreignKey  : fs.readFileSync(__dirname + '/foreign-key.sql').toString(),