+++ test/pg-structure-2.js
@@ -9,8 +9,31 @@ var db;
 
 
 describe('pg-structure', function () {
-    after(function (done) {
-        testDB.dropDB(done);
+    afterEach(function (done) {
+        //testDB.dropDB(done);
+        done();
+    });
+
+    it('should be ok with empty databases without any table.', function (done) {
+        testDB.resetDB(0, function () {
+            structure('localhost', 'pg_generator_test_724839', testDB.dbConfig.user, testDB.dbConfig.password, {schema: 'public'}, function (err, result) {
+                if (err) { throw err; }
+                db = result;
+                assert.equal(db.schema('public').name(), 'public');
+                done();
+            });
+        });
+    });
+
+    it('should be ok with databases with tables without any column.', function (done) {
+        testDB.resetDB('0b', function () {
+            structure('localhost', 'pg_generator_test_724839', testDB.dbConfig.user, testDB.dbConfig.password, {schema: 'public'}, function (err, result) {
+                if (err) { throw err; }
+                db = result;
+                assert.equal(db.schema('public').table('table_without_column').name(), 'table_without_column');
+                done();
+            });
+        });
     });
 
     it('should callback with informative error if accessed schema is not in DB and forgotten in options.', function (done) {
@@ -22,10 +45,16 @@ describe('pg-structure', function () {
             });
         });
     });
+});
 
+describe('pg-structure continued...', function () {
     it('should callback with error if parameters are not valid.', function (done) {
-        structure(99, 'pg_generator_test_724839', testDB.dbConfig.user, testDB.dbConfig.password, {schema: ['public']}, function (err, result) {
-            assert.throw(function () { if (err) { throw err; } }, /host must be a string/);
+        structure(99, 'pg_generator_test_724839', testDB.dbConfig.user, testDB.dbConfig.password, {schema: ['public']}, function (err) {
+            assert.throw(function () {
+                if (err) {
+                    throw err;
+                }
+            }, /host must be a string/);
             done();
         });
     });
@@ -42,5 +71,4 @@ describe('pg-structure', function () {
         }, /callback must be a Function/);
         done();
     });
-
 });