+++ test/util/create-test-db-0b.sql
@@ -1,0 +1,17 @@
+/*
+Created: 02/10/2014
+Modified: 24/11/2014
+Project: Node Test
+Model: Node Test
+Company: Fortibase
+Author: Özüm Eldoğan
+Version: 1.0
+Database: PostgreSQL 9.2
+*/
+
+-- Database Schema With A Table Without Columns
+
+CREATE TABLE "table_without_column"
+(
+
+)